myProject/requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.4.2
defusedxml==0.5.0
Django==2.1.3
django-rest-swagger==2.1.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
djangorestframework-xml==1.3.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
jwt==0.5.4
MarkupSafe==1.1.0
openapi-codec==1.3.2
Pillow==5.3.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycparser==2.19
PyJWT==1.6.4
pytz==2018.7
requests==2.20.1
simplejson==3.16.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.24.1
django-environ==0.4.5